chapter07/section07/src/main/java/cn/ep/spring/boot/ch07/s07/CustomConfig.java
package cn.ep.spring.boot.ch07.s07;

import org.springframework.beans.factory.annotation.Value;
import org.springframework.context.annotation.Configuration;

@Configuration
public class CustomConfig {

    @Value("${prop.key:1}")
    private String key;
    @Value("${prop.value:2}")
    private String value;

    public String getKey() {
        return key;
    }

    public void setKey(String key) {
        this.key = key;
    }

    public String getValue() {
        return value;
    }

    public void setValue(String value) {
        this.value = value;
    }
}
